Add files via upload
Welcome to the repository for my Netflix homepage clone! This project is a representation of the Netflix landing page, crafted with CSS for structure and for styling. The goal is to showcase not only the structure of the page but also the styling techniques used to replicate the visually appealing Netflix interface.